# Use AutoDiffScalar gradients in mex RBM method calls