--- a/drake/systems/controllers/bodyMotionControlmex.cpp
+++ b/drake/systems/controllers/bodyMotionControlmex.cpp
@@ -72,12 +72,12 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
   assert(mxGetM(prhs[narg])==6); assert(mxGetN(prhs[narg])==1);
   Map< Vector6d > body_vdot_des(mxGetPrSafe(prhs[narg++]));
 
-  KinematicsCache<double> cache = pdata->r->doKinematics(q, qd, 0); // FIXME: pass this into the mex function instead
+  KinematicsCache<double> cache = pdata->r->doKinematics(q, qd); // FIXME: pass this into the mex function instead
 
   // TODO: this must be updated to use quaternions/spatial velocity
   auto point = Vector3d::Zero().eval();
   auto body_pose = pdata->r->forwardKin(cache, point, pdata->body_index, 0, 1);
-  auto J = pdata->r->forwardKinJacobian(cache, point, pdata->body_index, 0, 1, false, 0).value();
+  auto J = pdata->r->forwardKinJacobian(cache, point, pdata->body_index, 0, 1, false);
 
   Vector6d body_error;
   body_error.head<3>()= body_pose_des.head<3>()-body_pose.head<3>();