--- a/drake/systems/controllers/controlUtil.cpp
+++ b/drake/systems/controllers/controlUtil.cpp
@@ -110,7 +110,7 @@ int contactConstraintsBV(RigidBodyManipulator *r, const KinematicsCache<double>&
     if (nc>0) {
       for (auto pt_iter=iter->contact_pts.begin(); pt_iter!=iter->contact_pts.end(); pt_iter++) {
         contact_pos = r->forwardKin(cache, *pt_iter, iter->body_idx, 0, 0);
-        J = r->forwardKinJacobian(cache, *pt_iter, iter->body_idx, 0, 0, true, 0).value();
+        J = r->forwardKinJacobian(cache, *pt_iter, iter->body_idx, 0, 0, true);
 
         normal = iter->support_surface.head(3);
         surfaceTangents(normal,d);
@@ -126,8 +126,7 @@ int contactConstraintsBV(RigidBodyManipulator *r, const KinematicsCache<double>&
         // NOTE: I'm cheating and using a slightly different ordering of J and Jdot here
         Jp.block(3*k,0,3,nq) = J;
         Vector3d pt = (*pt_iter).head(3);
-        auto Jpdotv_grad = r->forwardJacDotTimesV(cache, pt,iter->body_idx,0,0,0);
-        Jpdotv.block(3*k,0,3,1) = Jpdotv_grad.value();
+        Jpdotv.block(3*k,0,3,1) = r->forwardJacDotTimesV(cache, pt,iter->body_idx,0,0);
         normals.col(k) = normal;
         
         k++;
@@ -282,13 +281,13 @@ Vector6d bodySpatialMotionPD(RigidBodyManipulator *r, DrakeRobotState &robot_sta
   Vector3d body_xyz_task = T_world_to_task * body_xyz.colwise().homogeneous();
   Vector4d body_quat = body_pose.tail<4>();
   std::vector<int> v_indices;
-  auto J_geometric = r->geometricJacobian(cache, 0, body_index, body_index, 0, true, &v_indices);
+  auto J_geometric = r->geometricJacobian(cache, 0, body_index, body_index, true, &v_indices);
   VectorXd v_compact(v_indices.size());
   for(size_t i = 0;i<v_indices.size();i++)
   {
     v_compact(i) = robot_state.qd(v_indices[i]);
   }
-  Vector6d body_twist = J_geometric.value() * v_compact;
+  Vector6d body_twist = J_geometric * v_compact;
   Matrix3d R_body_to_world = quat2rotmat(body_quat);
   Matrix3d R_world_to_body = R_body_to_world.transpose();
   Matrix3d R_body_to_task = T_world_to_task.linear() * R_body_to_world;