--- a/drake/systems/plants/@RigidBodyManipulator/centroidalMomentumMatrixDotTimesV.m
+++ b/drake/systems/plants/@RigidBodyManipulator/centroidalMomentumMatrixDotTimesV.m
@@ -10,11 +10,17 @@ if (kinsol.mex)
   if (obj.mex_model_ptr==0)
     error('Drake:RigidBodyManipulator:InvalidKinematics','This kinsol is no longer valid because the mex model ptr has been deleted.');
   end
+  Adot_times_v = centroidalMomentumMatrixDotTimesVmex(obj.mex_model_ptr, kinsol.mex_ptr, 0, robotnum);
   if compute_gradients
-    [Adot_times_v, dAdot_times_v] = centroidalMomentumMatrixDotTimesVmex(obj.mex_model_ptr, kinsol.mex_ptr, 1, robotnum);
-  else
-    Adot_times_v = centroidalMomentumMatrixDotTimesVmex(obj.mex_model_ptr, kinsol.mex_ptr, 0, robotnum);
+    [Adot_times_v, dAdot_times_v] = eval(Adot_times_v);
+    nq = length(kinsol.q);
+    if isempty(dAdot_times_v)
+      dAdot_times_v = double.empty(0, nq);
+    else
+      dAdot_times_v = dAdot_times_v(:, 1 : nq);
+    end
   end
+  
 else
   
   nq = obj.getNumPositions();