--- a/drake/systems/plants/@RigidBodyManipulator/centroidalMomentumMatrixDotTimesV.m
+++ b/drake/systems/plants/@RigidBodyManipulator/centroidalMomentumMatrixDotTimesV.m
@@ -11,12 +11,11 @@ if (kinsol.mex)
     error('Drake:RigidBodyManipulator:InvalidKinematics','This kinsol is no longer valid because the mex model ptr has been deleted.');
   end
   Adot_times_v = centroidalMomentumMatrixDotTimesVmex(obj.mex_model_ptr, kinsol.mex_ptr, 0, robotnum);
-  if compute_gradients
+  if kinsol.has_gradients
     [Adot_times_v, dAdot_times_v] = eval(Adot_times_v);
     nq = length(kinsol.q);
-    dAdot_times_v = dAdot_times_v(:, 1 : nq);
+    dAdot_times_v = dAdot_times_v(:, 1 : nq); % for backwards compatibility
   end
-  
 else
   nq = obj.getNumPositions();
   if compute_gradients