Instead of using user gradients, use AutoDiffScalar gradients in all relevant RBM mex functions.

Note that the current implementation will try to call method instantiations with a fixed maximum size derivative vector (with a max size up to 73, the number of states of a quat-parameterized Atlas) if possible, before gracefully falling back to true dynamic size derivative vectors. See https://github.com/RobotLocomotion/drake/issues/1199 for the performance comparison that inspired this choice.

I made it so that TaylorVar inputs to the functions also call mex functions.

In the process:
* added some more functionality to mexify:
 * added mexTryToCallFunctions, a function to which you can pass a bunch of std::functions, which will be tried in order until one sticks in the sense that the Matlab arguments passed in can be converted to the correct types required by the function ~~without a MexToCPPConversionError being thrown~~. I'm using this to switch between the three current options of double, auto diff fixed max size, and auto diff dynamic max size
 * a small infrastructure improvement regarding the number of outputs of the function
 * mex conversions to general std::vectors and std::pairs
 * improved conversion to AutoDiffScalar
 * moved conversion functions to standardMexConversions.h and rigidBodyManipulatorMexConversions.h
 * added some documentation
* mexified contactConstraintsmex
* fixed some more 'scalar correctness' issues in RBM (places where we were wrongly assuming doubles)
* implemented a (necessarily imperfect) make_function function that works for all of our use cases so far, and use it to clean up rbmMexFunctions
* fixed an issue in the gradients of FixedAxisOneDoFJoint frictionTorque gradients that was present both in Matlab and in C++
* added a Matlab sparse matrix to Eigen sparse conversion

~~This isn't quite ready to go in yet, but I wanted to get test results and to get a discussion started:~~
Ready to go after review and build servers say OK (MSVC might have trouble again).

1. ~~the mexTryToCallFunctions approach looks like it may have too much overhead when gradient output is required. I'll do some more benchmarking to make sure. Any ideas on clean alternatives or ways to improve performance?~~ **rewrote mexTryToCallFunctions without try/catch; much less overhead.**
1. The current return type of AutoDiffScalar methods is a TaylorVar matrix. Is that a good idea, or should I convert to a series of matrices? **returning TaylorVar is quite expensive due to the mexCallMatlabSafe call. Should probably return a bunch of matrices instead, but in a future iteration. Not a major concern.**
1. ~~This currently increases compilation time quite a bit due to the two extra sets of explicit RBM method instantiations. We could just completely switch to AutoDiff gradients and get rid of the user gradients to offset this, but I didn't want to make this pull request too big before discussing. What do you think?~~ **got rid of user gradients**
1. ~~More of a note to self: I have to make sure that all of the gradient tests still make sense, since gradients obtained by requesting more output arguments for a function and TaylorVar gradients are now computed by calling the exact same code.~~ **seems to be OK. Tests either don't use mex for TaylorVar gradient checks, or they use numerical gradients.**