@tkoolen 2015-10-19
Internal compiler error on MSVC.