@tkoolen 2015-10-20
MSVC: `fatal error C1060: compiler is out of heap space` while compiling RBM... Might need to get rid of the user gradients or find a different solution before this can be accepted.