@tkoolen 2015-10-22
Got rid of the user gradients in RigidBodyManipulator. There are still some lower level things that use GradientVar, which I will clean up later.