@tkoolen 2015-10-26
Updated description above.