@tkoolen 2015-10-27
Do we care about the testIK timeout on drake006?