@RussTedrake 2015-10-27
i do care.  but it's a known (to me) issue, not the fault of your pull 
request

> Twan Koolen <mailto:notifications@github.com>
> October 27, 2015 at 1:59 PMvia Postbox 
> <https://www.postbox-inc.com/?utm_source=email&utm_medium=sumlink&utm_campaign=reach>
>
> Do we care about the testIK timeout on drake006?
>
> —
> Reply to this email directly or view it on GitHub 
> <https://github.com/RobotLocomotion/drake/pull/1434#issuecomment-151587872>.
>