@RussTedrake 2015-11-21
@tkoolen -- it looks amazing.  thank you for doing all this!  
i think you still have one legitimate test failing.  and I worry about big changes in performance -- have you done the benchmarking?
otherwise, to my (now very tired) eyes, it looks good to go. :)