@tkoolen 2015-11-21
I'll take a look at the failing test.

As for performance, I think I did a pretty decent job of assessing the performance of the C++ code, see https://github.com/RobotLocomotion/drake/issues/1199 and https://github.com/RobotLocomotion/drake/pull/1350. My assumption here has been that forwardKinJacobian, massMatrix and dynamicsBiasTerm are representative methods.

I also benchmarked the overhead of the mexTryToCallFunctions approach and deemed it acceptable (after moving away from the try/catch approach).

I'll do one more assessment by running `benchmarkRigidBodyManipulator.m` both before and after these changes to make sure that I didn't introduce too much overhead on the Matlab side of things.

Thank you for the review!