@tkoolen 2015-11-23
Addressed comments.

Benchmark results (benchmarkRigidBodyManipulator.m, times in ms per call):

Scenario 1: a couple of Jacobians

 | before | after
--- | --- | ---
no gradients | 0.5642 | 0.6280
with gradients | 1.549 | 2.546

Scenario 2: manipulatorDynamics

 | before | after
--- | --- | ---
no gradients | 0.2329 | 0.2681
with gradients | 3.924 | 3.077

There's quite a bit of overhead constructing and eval'ing TaylorVars, so I think we can get the gradient times down significantly by just using structs with f and df fields as the mex interface. In my opinion, that can wait until a later pull request though.

forwardKinmex has become quite a bit slower it seems:

Scenario 1

 | before | after
--- | --- | ---
forwardKinJacobianmex | 0.824 | 0.895
forwardKinmex | 0.156 | 0.564
doKinematicsmex | 0.360 | 0.317

Instruments showed that much of the time difference in forwardKinmex and forwardKinJacobianmex is attributable to checking that we're dealing with the right Scalar type in `KinematicsCache<Scalar>', which is currently done by doing a string compare between the output of `typeid' and the name stored in the DrakeMexPointer. This could definitely be improved a lot.