@tkoolen 2015-11-23
Yeah sure, I can fix it. My idea is to have an unique integer for each C++ type that we'll use in a DrakeMexPointer, and to store that integer as an additional field in DrakeMexPointer.