@tkoolen 2015-11-24
Some optimizations; effects turned out to be quite minor. See updated numbers in the previous post.

I found out why forwardKinmex seemed so slow: the 'after' included calls to forwardKinmex with TaylorVar inputs; comparing apples to oranges... So discard that and just look at the first two tables.

I think the overhead is still ridiculous (without gradients, only 11% of forwardKinmex time is spent in RigidBodyManipulator::forwardKin, 32% for forwardKinJacobianmex, 35% for dynamicsBiasTermmex, 59% for massMatrixmex), but it turned out to be hard to get rid of. Specifically mxGetProperty calls are very expensive and needed to process the DrakeMexPointers.

For scenario 1, I think the old gradients were faster because I actually implemented something better than just taking gradients of every operation in the algorithm (pretty much the only case of that).