@RussTedrake 2015-11-24
interesting.  not asking for this now, but it should still be possible to implement user gradients (by populating the autodiff structures manually) if we choose... correct?  Perhaps it's worth adding a comment now in the code that used to be more optimized, with a link to a revision where we could find that code if we want to improve it later?