@tkoolen 2015-11-25
> interesting. not asking for this now, but it should still be possible to implement user gradients (by populating the autodiff structures manually) if we choose... correct?

Yeah, you can override the method for AutoDiffScalar inputs and directly set the derivatives if you want. I'll put it in the 'rbm cleanup list' issue.