@avalenzu 2015-11-26
Wow! Congrats, Twan!

On Thu, Nov 26, 2015, 4:06 AM Russ Tedrake <notifications@github.com> wrote:

> Merged #1434 <https://github.com/RobotLocomotion/drake/pull/1434>.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/RobotLocomotion/drake/pull/1434#event-475365664>.
>